requirements.txt
django==5.0.6
djangorestframework==3.15.2
django-cors-headers
psycopg2-binary
oauth2_provider
djangorestframework-simplejwt
django-oauth-toolkit
Pillow
opencv-python-headless
olefile
uuid
